// __generated__/graphql.ts
export enum Episode {
  Empire = 'EMPIRE',
  Jedi = 'JEDI',
  NewHope = 'NEW_HOPE'
}